% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_functions.R
\name{fake_correlation}
\alias{fake_correlation}
\title{Generate Fake Correlation Data}
\usage{
fake_correlation(
  n = 100,
  correlation = 0.8,
  mean_x = 0,
  mean_y = 0,
  sd_x = 1,
  sd_y = 1
)
}
\arguments{
\item{n}{Number of data points to generate}

\item{correlation}{Desired correlation coefficient (between -1 and 1)}

\item{mean_x}{Mean of the x variable}

\item{mean_y}{Mean of the y variable}

\item{sd_x}{Standard deviation of the x variable}

\item{sd_y}{Standard deviation of the y variable}
}
\value{
A data frame with x and y variables showing the desired correlation
}
\description{
This function generates fake data that will always show the correlation
you want to see, regardless of what your actual data looks like.
Perfect for when reality doesn't match your hypothesis!
}
\examples{
# Generate data with perfect positive correlation
fake_data <- fake_correlation(n = 100, correlation = 0.95)
plot(fake_data$x, fake_data$y)
cor(fake_data$x, fake_data$y)

# Generate data with strong negative correlation
fake_data2 <- fake_correlation(n = 50, correlation = -0.8)
cor(fake_data2$x, fake_data2$y)
}
